requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
dominate==2.7.0
Flask==2.2.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
idna==3.4
importlib-metadata==6.5.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pydantic==1.10.7
PyJWT==2.6.0
pytz==2023.3
requests==2.28.2
StrEnum==0.4.10
typing_extensions==4.5.0
urllib3==1.26.15
visitor==0.1.3
waitress==2.1.2
Werkzeug==2.2.3
WTForms==3.0.1
zipp==3.15.0
langchain==0.0.142
openai~=0.27.5
faiss-cpu
streamlit~=1.22.0
pypdf
streamlit-chat
huggingface_hub
llama-index
streamlit-analytics
watchdog
pinecone-client
unstructured
streamlit_option_menu
bs4
embedchain~=0.0.10
docx~=0.2.4
